提交api管理分页查询以及不分页查询接口

+++ goSystemManage/goAdmin/base/request/sys_base.go
@@ -1,7 +1,7 @@
 package request
 
 type GetById struct {
-	Id float64 `json:"id" form:"id"`
+	Id int `json:"id" form:"id"`
 }
 
 type GetIds struct {
